refactor: reorganize modules and add executor
