+++ sigma/core/strategies.py
