+++ sigma/utils/logger.py
@@ -1,0 +1,3 @@
+from loguru import logger
+
+__all__ = ["logger"]
